refactor(illustration-exhibit-details): extract download functions to own utils file
