+++ packages/web/src/app/doc-pages/brand/illustrations-doc/illustrations-exhibit/illustrations-exhibit-details/illustrations-exhibit-details.component.ts
@@ -3,6 +3,7 @@ import { toSignal } from '@angular/core/rxjs-interop';
 
 import { IllustrationsExhibitService } from '../illustrations-exhibit.service';
 import { IllustrationsGeneratorComponent } from '../illustrations-generator/illustrations-generator.component';
+import { createPngBlob, createSvgBlobFromElement } from './imageDownloadUtils';
 import { LocalizationService } from 'src/app/core/services/localization.service';
 import { CopyComponent } from 'src/app/shared/copy/copy.component';
 
@@ -51,76 +52,25 @@ export class IllustrationsExhibitDetailsComponent {
 
   downloadSvg = () => {
     const a = document.createElement('a');
-    a.href = this.createSvgBlobFromElement();
+    a.href = createSvgBlobFromElement(this.getIllustrationElement(), this.colorValue());
     a.download = this.imageFileName('svg');
     a.click();
   };
 
   downloadPng = async () => {
     const a = document.createElement('a');
-    a.href = await this.createPngBlob();
+    a.href = await createPngBlob(this.getIllustrationElement(), this.colorValue());
     a.download = this.imageFileName('png');
     a.click();
   };
 
-  private createSvgBlobFromElement = () => {
-    const selectedIllustration = this.selectedIllustration();
-    if (!selectedIllustration) {
-      return '';
-    }
-
-    const illustrationElement = (this.elementRef.nativeElement as HTMLElement).querySelector(
-      `elvia-illustrations-${selectedIllustration}`,
+  private getIllustrationElement = () => {
+    const element = (this.elementRef.nativeElement as HTMLElement).querySelector(
+      `elvia-illustrations-${this.selectedIllustration()}`,
     );
-    const svgElement = illustrationElement?.shadowRoot?.querySelector('svg');
-    const styleElement = illustrationElement?.shadowRoot?.querySelector('style');
-    if (!styleElement || !svgElement) {
-      console.error('No SVG or style element found');
-      return '';
+    if (!element) {
+      throw new Error('No illustration element found');
     }
-
-    const newSvgElement = svgElement.cloneNode(true) as SVGElement;
-    newSvgElement.insertBefore(styleElement.cloneNode(true), newSvgElement.firstChild);
-    newSvgElement.classList.add(this.colorValue() || '');
-
-    const svgString = new XMLSerializer().serializeToString(newSvgElement);
-    const blob = new Blob([svgString], { type: 'image/svg+xml' });
-    return URL.createObjectURL(blob);
-  };
-
-  private createPngBlob = (): Promise<string> => {
-    return new Promise((resolve, reject) => {
-      const svgBlobUrl = this.createSvgBlobFromElement();
-
-      if (!svgBlobUrl) {
-        console.error('No SVG blob url found');
-        reject(new Error('No SVG blob url found'));
-      }
-
-      const img = new Image();
-      img.src = svgBlobUrl;
-      img.onload = () => {
-        const canvas = document.createElement('canvas');
-        const ctx = canvas.getContext('2d');
-        if (!ctx) {
-          console.error('Could not get 2d context');
-          reject(new Error('Could not get 2d context'));
-          return;
-        }
-        canvas.width = 8 * img.width;
-        canvas.height = 8 * img.height;
-        ctx.drawImage(img, 0, 0);
-
-        canvas.toBlob((blob) => {
-          if (!blob) {
-            console.error('Could not create blob');
-            reject(new Error('Could not create blob'));
-            return;
-          }
-
-          resolve(URL.createObjectURL(blob));
-        });
-      };
-    });
+    return element;
   };
 }
